% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timenow.R
\name{detect_timezone}
\alias{detect_timezone}
\title{Detect the local timezone}
\usage{
detect_timezone(quiet = FALSE)
}
\arguments{
\item{quiet}{Suppress messages about detection method used}
}
\value{
A valid IANA timezone string
}
\description{
Uses a cascade of methods to determine the local timezone, useful when
the system timezone is set to UTC.
}
